models/BiSeNet/requirements.txt
absl-py==0.11.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
dataclasses==0.6
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
importlib-metadata==2.0.0
kiwisolver==1.3.0
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.4.0.44
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rich==9.1.0
rsa==4.6
six==1.15.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
zipp==3.4.0
